+++ src/components/avatar-button/avatar-button.tsx
@@ -1,21 +1,16 @@
 import { FC } from "react";
-import { Avatar, AvatarProps } from "../avatar";
+import { Avatar, AvatarProps } from "@/components/avatar";
 
-type AvatarButtonProps = {
-  avatarProps: AvatarProps;
-  onClick?: () => void;
-};
+type AvatarButtonProps = AvatarProps;
 
 export const AvatarButton: FC<AvatarButtonProps> = ({
-  avatarProps,
-  onClick,
-}) => {
-  return (
-    <button
-      onClick={onClick}
-      className="p-[8px] cursor-pointer flex justify-center items-center bg-primary-600 hover:bg-primary-700 flex-col rounded-lg"
-    >
-      <Avatar {...avatarProps} size="s" />
-    </button>
-  );
-};
+  fallbackText,
+  ...avatarProps
+}) => (
+  <div
+    aria-label={fallbackText ?? "Avatar"}
+    className="p-2 cursor-pointer flex justify-center items-center bg-primary-600 hover:bg-primary-700 flex-col rounded-lg focus-ring-neutral"
+  >
+    <Avatar {...avatarProps} fallbackText={fallbackText} size="s" />
+  </div>
+);