+++ src/components/avatar-button/tests/avatar-button.unit.test.tsx
@@ -1,17 +1,20 @@
-import { render, screen, fireEvent } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import { describe, it, expect, vi } from "vitest";
+import { describe, it, expect } from "vitest";
 import { AvatarButton } from "../index";
 
 describe("AvatarButton", () => {
   it("renders avatar correctly", () => {
     render(
       <AvatarButton
-        avatarProps={{
-          src: "/avatars/rory.jpg",
-          alt: "User Avatar",
-          fallbackText: "Fallback",
-        }}
+        imageElement={
+          <img
+            src="/avatars/rory.jpg"
+            alt="User Avatar"
+            className="object-cover w-full h-full"
+          />
+        }
+        fallbackText="Fallback"
       />,
     );
     expect(screen.getByTestId("avatar-root")).toBeInTheDocument();
@@ -20,32 +23,19 @@ describe("AvatarButton", () => {
   it("overrides avatar size prop", () => {
     render(
       <AvatarButton
-        avatarProps={{
-          src: "/avatars/rory.jpg",
-          alt: "User Avatar",
-          fallbackText: "Fallback",
-          size: "xl",
-        }}
+        imageElement={
+          <img
+            src="/avatars/rory.jpg"
+            alt="User Avatar"
+            className="object-cover w-full h-full"
+          />
+        }
+        fallbackText="Fallback"
+        size="xl"
       />,
     );
     const avatar = screen.getByTestId("avatar-root");
     expect(avatar.className).toContain("w-10");
     expect(avatar.className).toContain("h-10");
-  });
-
-  it("calls onClick handler when button is clicked", () => {
-    const handleClick = vi.fn();
-    render(
-      <AvatarButton
-        avatarProps={{
-          src: "/avatars/rory.jpg",
-          alt: "User Avatar",
-          fallbackText: "Fallback",
-        }}
-        onClick={handleClick}
-      />,
-    );
-    fireEvent.click(screen.getByRole("button"));
-    expect(handleClick).toHaveBeenCalledTimes(1);
   });
 });