+++ src/components/file-upload/components/upload-button.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
-import { UploadIcon } from "../../icons";
-import { Button } from "../../button";
+import { UploadIcon } from "@/components/icons";
+import { Button } from "@/components/button";
 
 type UploadButtonProps = {
   label: string;
@@ -8,7 +8,13 @@ type UploadButtonProps = {
 };
 
 export const UploadButton: FC<UploadButtonProps> = ({ label, onClick }) => (
-  <Button variant="light" size="default" fullWidth onClick={onClick}>
+  <Button
+    type="button"
+    variant="light"
+    size="default"
+    fullWidth
+    onClick={onClick}
+  >
     <span className="mr-2">{label}</span>
     <UploadIcon color="neutral" />
   </Button>