+++ src/compositions/index.ts
@@ -1,0 +1,6 @@
+export * from "./app-shell";
+export * from "./header";
+export * from "./upload-image-modal";
+export * from "./profile-header";
+export * from "./user-card";
+export * from "./user-info";